requirements.txt
discord.py
PyDictionary
Wikipedia-API
youtube-search
asyncpg